# backend/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "成果云系统API文档",
    "description": "成果云系统后端接口文档（Flask + MySQL）",
    "version": "1.0.0"
  },
  "basePath": "/",
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "登录后返回的 Bearer Token，格式：Bearer <access_token>"
    }
  },
...
  "paths": {
...
    "/api/teachers/": {
      "get": {
        "tags": ["教师"],
        "summary": "教师列表（分页）",
        "security": [{"BearerAuth": []}],
        "responses": {"200": {"description": "返回教师列表"}}
      },
      "post": {
        "tags": ["教师"],
        "summary": "添加教师",
        "security": [{"BearerAuth": []}],
        "responses": {"201": {"description": "添加成功"}}
      }
    },
    "/api/teachers/statistics": {
      "get": {
        "tags": ["教师"],
        "summary": "教师统计（总数/待审核）",
        "security": [{"BearerAuth": []}],
        "responses": {"200": {"description": "返回统计数字"}}
      }
    },
    "/api/teachers/{id}": {
      "parameters": [
        {"name": "id", "in": "path", "required": true, "type": "integer"}
      ],
      "get": {
        "tags": ["教师"],
        "summary": "教师详情",
        "security": [{"BearerAuth": []}],
        "responses": {"200": {"description": "返回教师详情"}}
      },
      "put": {
        "tags": ["教师"],
        "summary": "编辑教师",
        "security": [{"BearerAuth": []}],
        "responses": {"200": {"description": "更新成功"}}
      },
      "delete": {
        "tags": ["教师"],
        "summary": "删除教师",
        "security": [{"BearerAuth": []}],
        "responses": {"200": {"description": "删除成功"}}
      }
    },
    "/api/teachers/{id}/approve": {
      "post": {
        "tags": ["教师"],
        "summary": "审核教师",
        "security": [{"BearerAuth": []}],
        "responses": {"200": {"description": "审核成功"}}
      }
    },
...
  }
}